split relations module
